test & docs: добавление новых тестов для БД и их документирование
